Added service for SendGrid V3 API.
Added a service for the new SendGrid V3 API. If you configure
`nxus-mailer` with an `apiKey`, instead of `apiUsername` and
`apiPassword`, the old SendGrid service won't start, but the new one
will. (Open to just ditching the old service, too.)

The new `send()` method checks whether the message content is HTML or
plain text, and configures the data for SendGrid appropriately. If the
content is HTML, it generates a plain text equivalent.

The `send()` method also recognizes an `apiKey` option, to override the
API Key on a per-message basis. (I'm thinking gf-assess, sprawling mess
that it is, may want to use multiple API keys – to distinguish password
reset requests from policy page contacts, for example.)

Also updated to Babel 7 and switched tests to Jest. Set Babel
`preset-env` to target a minimum node version of 8. (Does this seem
reasonable?) Updated documentation, too. (Can you believe
`documentation` has Babel dependencies?)

Are we adding `package-lock.json` to modules now?